[MISC] introduce snippet::alphabet_base::{rank_to_char, char_to_rank} function
